requirements.txt
tensorflow~=2.18.0
keras~=3.7.0
numpy
opencv-python~=4.10.0.84
pillow~=11.0.0